// src/api/models/UpdateUserBodyDto.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type UpdateUserBodyDto = {
    email?: string;
    roles?: Array<string>;
    permissions?: Array<string>;
    password?: string;
    /**
     * User ID
     */
    id: string;
};